fix: broken chat search on redis
Redisearch docs indicate that whatever is passed in is evaluated against index data.
Therefore, even if uid is passed in and is falsy, it is still evaluated and does not match any indexed data.

This commit updates the `roomId` and `uid` properties so that they must be arrays containing items, similarly to how the postgres method handles values.